Reorganize and clean up repository
